test(e2e): full-app smoke under test/ (headless); drop integration_test dep (#152)
Re-scoped during the spike: the `integration_test` package requires a device
(local `flutter test integration_test/` errors "more than one device"; the spike
only ran via `-d linux` with heavy native deps — libsecret/curl/GTK/sentry-native).
That makes "headless CI" untrue for integration_test/.

A full-app widget test under `test/` delivers the same e2e value (boot → login →
dashboard, exercising routing + DI + BLoCs + mock HTTP) and runs headless under
the existing `flutter test` with zero CI cost — `App` self-wraps and is already
pumpable headless (proven by app_test.dart). The global bootstrap supplies the
secure-storage mock + storage reset, so no per-file setup is needed.

- Add `test/integration/app_smoke_test.dart` (cold start → login(mock) → dashboard).
- Remove the unused `integration_test` dev dependency (we're not using the package).
- Docs (testing-architecture.md, README) describe the smoke + note that real
  on-device e2e would re-add integration_test + a device-targeted CI job.

Closes #152.
