+++ .gitignore
@@ -64,3 +64,7 @@ test_results/
 
 # Alchemist golden failure diffs
 test/**/failures/
+
+# Stray native-build artifacts (never commit)
+CMakeFiles/
+*.deb
