+++ pubspec.yaml
@@ -46,8 +46,6 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_lints: 6.0.0
-  integration_test:
-    sdk: flutter
   test: 1.31.0
   bloc_test: 10.0.0
   mocktail: 1.0.5
